allow custom object_hook in json.loads
in order to parse dates, a custom object_hook is required. the method for determining what is a datetime could vary from project to project. this likely would be best done by checking the 'type' field from a collections template, but that is too complex for now.

furthermore, there may be instances when a different object hook is desired, other than checking for datetimes.